:truck: rename: 보고서 request, response 폴더 정리
