+++ src/main/java/com/sponus/sponusbe/domain/report/controller/ReportController.java
@@ -13,11 +13,11 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.sponus.sponusbe.auth.annotation.AuthOrganization;
 import com.sponus.sponusbe.domain.organization.entity.Organization;
-import com.sponus.sponusbe.domain.report.dto.ReportCreateRequest;
-import com.sponus.sponusbe.domain.report.dto.ReportCreateResponse;
-import com.sponus.sponusbe.domain.report.dto.ReportGetResponse;
-import com.sponus.sponusbe.domain.report.dto.ReportUpdateRequest;
-import com.sponus.sponusbe.domain.report.dto.ReportUpdateResponse;
+import com.sponus.sponusbe.domain.report.dto.request.ReportCreateRequest;
+import com.sponus.sponusbe.domain.report.dto.response.ReportCreateResponse;
+import com.sponus.sponusbe.domain.report.dto.response.ReportGetResponse;
+import com.sponus.sponusbe.domain.report.dto.request.ReportUpdateRequest;
+import com.sponus.sponusbe.domain.report.dto.response.ReportUpdateResponse;
 import com.sponus.sponusbe.domain.report.service.ReportQueryService;
 import com.sponus.sponusbe.domain.report.service.ReportService;
 import com.sponus.sponusbe.global.common.ApiResponse;
