+++ src/main/java/com/sponus/sponusbe/domain/report/dto/response/ReportImageResponse.java
@@ -1,4 +1,4 @@
-package com.sponus.sponusbe.domain.report.dto;
+package com.sponus.sponusbe.domain.report.dto.response;
 
 import com.sponus.sponusbe.domain.report.entity.ReportImage;
