+++ src/main/java/com/sponus/sponusbe/domain/report/dto/response/ReportUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.sponus.sponusbe.domain.report.dto;
+package com.sponus.sponusbe.domain.report.dto.response;
 
 import com.sponus.sponusbe.domain.report.entity.Report;
