+++ src/main/java/com/sponus/sponusbe/domain/report/service/ReportQueryService.java
@@ -3,9 +3,7 @@ package com.sponus.sponusbe.domain.report.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.sponus.sponusbe.domain.organization.entity.Organization;
-import com.sponus.sponusbe.domain.report.dto.ReportCreateResponse;
-import com.sponus.sponusbe.domain.report.dto.ReportGetResponse;
+import com.sponus.sponusbe.domain.report.dto.response.ReportGetResponse;
 import com.sponus.sponusbe.domain.report.entity.Report;
 import com.sponus.sponusbe.domain.report.exception.ReportErrorCode;
 import com.sponus.sponusbe.domain.report.exception.ReportException;
